refactor(agents): split embedded tool result ownership
