+++ config/max-lines-baseline.txt
@@ -389,7 +389,6 @@ src/agents/embedded-agent-subscribe.handlers.messages.test.ts
 src/agents/embedded-agent-subscribe.handlers.messages.ts
 src/agents/embedded-agent-subscribe.handlers.tools.test.ts
 src/agents/embedded-agent-subscribe.subscribe-embedded-agent-session.subscribeembeddedagentsession.test.ts
-src/agents/embedded-agent-subscribe.tools.ts
 src/agents/embedded-agent-subscribe.ts
 src/agents/failover-error.test.ts
 src/agents/failover-error.ts
