+++ src/agents/embedded-agent-subscribe.handlers.tools.completion.ts
@@ -31,6 +31,11 @@ import {
   resolveMessageToolSourceReplyFinal,
 } from "./embedded-agent-message-tool-source-reply.js";
 import {
+  extractMessagingToolSend,
+  extractMessagingToolSendResult,
+  extractMessagingToolSourceReplyPayload,
+} from "./embedded-agent-messaging-extraction.js";
+import {
   isMessagingTool,
   isMessagingToolSendAction,
   isMessagingToolTargetEvidenceAction,
@@ -77,16 +82,14 @@ import type { ToolHandlerContext } from "./embedded-agent-subscribe.handlers.typ
 import {
   collectMessagingMediaUrlsFromRecord,
   collectMessagingMediaUrlsFromToolResult,
+} from "./embedded-agent-tool-media.js";
+import {
   capLiveExecResult,
-  extractMessagingToolSourceReplyPayload,
   extractToolErrorCode,
-  extractMessagingToolSend,
-  extractMessagingToolSendResult,
   extractToolErrorMessage,
-  isToolResultError,
   isToolResultTimedOut,
   sanitizeToolResult,
-} from "./embedded-agent-subscribe.tools.js";
+} from "./embedded-agent-tool-results.js";
 import { parseExecApprovalResultText } from "./exec-approval-result.js";
 import { readMcpAppChannelView } from "./mcp-ui-resource.js";
 import type { AgentEvent } from "./runtime/index.js";
@@ -96,6 +99,7 @@ import {
 } from "./tool-error-summary.js";
 import { resolveFileMutationToolName } from "./tool-mutation-names.js";
 import { normalizeToolPolicyName } from "./tool-policy.js";
+import { isToolResultError } from "./tool-result-error.js";
 import { cancelAskUserPromptDelivery } from "./tools/ask-user-tool.js";
 import { isAutomationsToolName } from "./tools/automations-tool-name.js";
