+++ src/agents/embedded-agent-subscribe.tools.test-support.ts
@@ -1,4 +1,4 @@
-import "./embedded-agent-subscribe.tools.js";
+import "./embedded-agent-tool-media.js";
 
 type EmbeddedSubscribeToolsTestApi = {
   isToolResultMediaTrusted(
