+++ src/agents/embedded-agent-subscribe.ts
@@ -52,12 +52,12 @@ import type {
   EmbeddedAgentSubscribeContext,
   EmbeddedAgentSubscribeState,
 } from "./embedded-agent-subscribe.handlers.types.js";
+import type { SubscribeEmbeddedAgentSessionParams } from "./embedded-agent-subscribe.types.js";
 import {
-  buildToolLifecycleErrorResult,
   extractToolResultMediaArtifact,
   filterToolResultMediaUrls,
-} from "./embedded-agent-subscribe.tools.js";
-import type { SubscribeEmbeddedAgentSessionParams } from "./embedded-agent-subscribe.types.js";
+} from "./embedded-agent-tool-media.js";
+import { buildToolLifecycleErrorResult } from "./embedded-agent-tool-results.js";
 import {
   createThinkingTagStreamState,
   stripDowngradedToolCallText,
